refactor. Change structure of project. now new_core is core, old core deleted
