+++ src/main/java/bot/TGBot.java
@@ -1,7 +1,7 @@
 package bot;
 
 import models.Message;
-import new_core.MessageHandler;
+import core.MessageHandler;
 import org.apache.log4j.Logger;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
