+++ src/main/java/core/MessageHandler.java
@@ -1,5 +1,6 @@
 package core;
 
+import core.service_handlers.handlers.*;
 import db.LoggedUsersRepository;
 import db.UserContextRepository;
 import models.Message;
@@ -9,34 +10,110 @@ import models.UserContext;
 import java.sql.SQLException;
 
 /**
- * Обработчик сообщений
+ * Класс, обрабатывающий сообщения, поступающие от пользователей из разных платформ, направляющий
+ * сообщение в нужный сервисный обработчике, в зависимости от контекста пользователя для последующей
+ * обработки
  */
 public class MessageHandler {
+    /** @see UserContextRepository */
+    private final UserContextRepository userContextRepository;
 
-    /** обработчик команд */
+    /** @see LoggedUsersRepository*/
+    private final LoggedUsersRepository loggedUsersRepository;
+
+    /** @see CommandHandler */
     private final CommandHandler commandHandler;
 
-    /** обработчик текста */
-    private final TextHandler textHandler;
+    /** @see HandlerEditUserService */
+    private final HandlerEditUserService handlerEditUserService;
 
-    /**
-     * @param commandHandler обработчик команд
-     * @param textHandler обработчик текста
-     */
-    public MessageHandler(CommandHandler commandHandler, TextHandler textHandler){
+    /** @see HandlerCreateOrderService */
+    private final HandlerCreateOrderService handlerCreateOrderService;
+
+    /** @see HandlerEditOrderService */
+    private final HandlerEditOrderService handlerEditOrderService;
+
+    /** @see HandlerCancelOrderService */
+    private final HandlerCancelOrderService handlerCancelOrderService;
+    private final HandlerAcceptOrderService handlerAcceptOrderService;
+    private final HandlerCloseOrderService handlerCloseOrderService;
+
+
+    /** Конструктор {@link MessageHandler MessageHandler}*/
+    public MessageHandler(
+            UserContextRepository userContextRepository,
+            LoggedUsersRepository loggedUsersRepository,
+            CommandHandler commandHandler,
+            HandlerEditUserService updateUserServiceHandler,
+            HandlerCreateOrderService handlerCreateOrderService,
+            HandlerEditOrderService handlerEditOrderService,
+            HandlerCancelOrderService handlerCancelOrderService,
+            HandlerAcceptOrderService handlerAcceptOrderService,
+            HandlerCloseOrderService handlerCloseOrderService
+    ) {
+        this.userContextRepository = userContextRepository;
+        this.loggedUsersRepository = loggedUsersRepository;
         this.commandHandler = commandHandler;
-        this.textHandler = textHandler;
+        this.handlerEditUserService = updateUserServiceHandler;
+        this.handlerCreateOrderService = handlerCreateOrderService;
+        this.handlerEditOrderService = handlerEditOrderService;
+        this.handlerCancelOrderService = handlerCancelOrderService;
+        this.handlerAcceptOrderService = handlerAcceptOrderService;
+        this.handlerCloseOrderService = handlerCloseOrderService;
     }
 
-    /**
-     * метод определяет сообщение как команду или как текст
-     * и отправляет в соответствующий обработчик: {@link CommandHandler} или {@link TextHandler}
-     * @return Текст, который сформирует соответствующий обработчик
+    /** Первичный метод обработки сообщения. Если у пользователя, отправившего сообщение есть
+     * контекст, направляет сообщение в соответствующий сервисный обработчик, если отправленное
+     * сообщение является командой, перенаправляет сообщение в обработчик команд. Иначе сообщает
+     * пользователю, что сообщение некорректно.
      */
-    public String handle(Message msg){
-        String text = msg.getText();
-        if (text.charAt(0)=='/')
-            return commandHandler.handle(msg);
-        return textHandler.handle(msg);
+    public void handle(Message msg) {
+        User user;
+        UserContext userContext;
+        try {
+            user = loggedUsersRepository.getUserByPlatformAndIdOnPlatform(
+                    msg.getPlatform(),
+                    msg.getUserIdOnPlatform()
+            );
+        } catch (SQLException e) {
+            msg.getBotFrom().sendTextMessage(
+                    msg.getUserIdOnPlatform(),
+                    "Проблемы с доступом к базе данных " + e.getMessage()
+            );
+            return;
+        }
+
+        if(user == null) {
+            if (msg.getText().equals("/start")) {
+                commandHandler.handle(msg);
+                return;
+            }
+            msg.getBotFrom().sendTextMessage(
+                    msg.getUserIdOnPlatform(),
+                    "отправьте /start для последующей работы."
+            );
+            return;
+        }
+
+        msg.setUser(user);
+        try {userContext = userContextRepository.getUserContext(user.getId());}
+        catch (SQLException e) {
+            msg.getBotFrom().sendTextMessage(
+                    msg.getUserIdOnPlatform(),
+                    "Проблемы с доступом к базе данных " + e.getMessage());
+            return;
+        }
+
+        msg.setUserContext(userContext);
+        switch(userContext.getState()) {
+            case NO_STATE -> commandHandler.handle(msg);
+            case EDIT_USER -> handlerEditUserService.handle(msg);
+            case ORDER_CREATING -> handlerCreateOrderService.handle(msg);
+            case ORDER_EDITING -> handlerEditOrderService.handle(msg);
+            case ORDER_CANCELING-> handlerCancelOrderService.handle(msg);
+            case ORDER_ACCEPTING -> handlerAcceptOrderService.handle(msg);
+            case ORDER_CLOSING_COURIER,
+                    ORDER_CLOSING_CLIENT -> handlerCloseOrderService.handle(msg);
+        }
     }
 }
