+++ src/main/java/core/service_handlers/handlers/HandlerAcceptOrderService.java
@@ -1,8 +1,8 @@
-package new_core.service_handlers.handlers;
+package core.service_handlers.handlers;
 
 import models.Message;
-import new_core.service_handlers.services.AcceptOrderService;
-import new_core.service_handlers.services.CloseOrderService;
+import core.service_handlers.services.AcceptOrderService;
+import core.service_handlers.services.CloseOrderService;
 
 /** Обработчик контекстов {@link models.UserState#ORDER_ACCEPTING ORDER_ACCEPTING}*/
 public class HandlerAcceptOrderService {
