+++ src/main/java/core/service_handlers/handlers/HandlerCancelOrderService.java
@@ -1,8 +1,7 @@
-package new_core.service_handlers.handlers;
+package core.service_handlers.handlers;
 
 import models.Message;
-import new_core.service_handlers.services.CancelOrderService;
-import new_core.service_handlers.services.EditUserService;
+import core.service_handlers.services.CancelOrderService;
 
 /** Обработчик контекста {@link models.UserState#ORDER_CANCELING ORDER_CANCELING}. */
 public class HandlerCancelOrderService {
