+++ src/main/java/core/service_handlers/services/CloseOrderService.java
@@ -1,4 +1,4 @@
-package new_core.service_handlers.services;
+package core.service_handlers.services;
 
 import db.OrderRepository;
 import db.UserContextRepository;
